Rotated (ruleset, pointset, stateset).

+++ #/Automaton.2.cs
@@ -1,6 +1,6 @@
 partial class Automaton
 	{
-	const  int __default = 0 ;
+	const  ulong __default = 0 ;
 	static int[] xo_t ;
 	static int[] xo_l ;
 	static System.Action<Automaton>[] xo_a ;
